Extract docker-compose/Dockerfile template generation into container_compose.rs
Step 3 (the biggest chunk) of splitting containers.rs. Moves
site_hostnames, site_document_root, php_dockerfile, compose (kept
#[allow(clippy::too_many_arguments)]), and apply_runtime_defaults —
one large inline YAML-string builder with no prior sub-helpers. All
four became pub(crate) since site_hostnames is needed by ensure_gateway
(staying in containers.rs for now, moving in a later gateway step) in
addition to prepare(); no external re-export needed since nothing
outside containers.rs called these directly. The ~10 tests that
primarily exercise compose/php_dockerfile/site_document_root moved
with them into container_compose.rs's own test module (duplicating
the small test_environment()/test_site() fixtures there); the
remaining validate()-only, bootstrap, gateway, and CRUD tests stay in
containers.rs unchanged. 2882 -> 2201 lines (container_compose.rs:
770).
